Add unixtime function for GNU date
It prints the current unix time.

When a list of arguments is passed it does convert them from unix time
to an ISO date or from an (ISO but not necessary) date to unixtime,
depending on the arguments' format.

unixtime
--> 1471450882

Notice that it ignores any colon in the timezone.

unixtime 2016-08-17T18:21:22+02:00 2016-08-17T18:21:22+0200 1471450885
--> 1471450882
    1471450882
    2016-08-17T18:21:25+0200